+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 0.3.5
+- include `indiecert-housekeeping` script to remove expired approvals
+  where the user never came back and codes that were never claimed
+
 ## 0.3.4
 - update `fkooman/rest`
 - update `fkooman/rest-plugin-indieauth`